fixed an issue with unsetting environment value
